Backend/requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blis==0.7.11
boto3==1.34.134
botocore==1.34.134
breadability==0.1.20
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpathlib==0.18.1
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
docopt==0.6.2
drf-yasg==1.21.7
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
fastjsonschema==2.20.0
filelock==3.15.3
firebase-admin==6.5.0
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gcloud==0.18.3
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==22.10.0
inflection==0.5.1
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jwcrypto==1.5.6
kiwisolver==1.4.5
kubernetes==30.1.0
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
llama-cloud==0.0.6
llama-index==0.10.52
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.52.post1
llama-index-embeddings-cohere==0.1.9
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.3
llama-index-legacy==0.9.48
llama-index-llms-cohere==0.2.0
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.27
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-chroma==0.1.10
llama-parse==0.4.5
loguru==0.7.2
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.10
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parsel==1.9.1
parso==0.8.4
pdfminer.six==20231228
pdftitle==0.11
pillow==10.3.0
platformdirs==4.2.2
posthog==3.5.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.3.1
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.8.0
pymongo==3.12.3
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
Pyrebase4==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python-jwt==4.1.0
python-multipart==0.0.9
python-snippets==0.0.51
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
queuelib==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
Scrapy==2.11.2
selenium==4.21.0
Send2Trash==1.8.3
sentencepiece==0.2.0
service-identity==24.1.0
setuptools==70.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
sqlparse==0.2.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sumy==0.11.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.5.0
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.19.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.42.1
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.1
w3lib==2.2.1
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
whitenoise==6.7.0
win32-setctime==1.1.0
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.0
yarl==1.9.4
zipp==3.19.2
zope.interface==6.4.post2
